Minor bug fixes in data packages + improved info schema demo to connect to mysql and postgres as well as sql server..

+++ Extenso.Data.Npgsql/NpgsqlConnectionExtensions.cs
@@ -150,7 +150,7 @@ AND kcu.""table_schema"" = @SchemaName";
                             //}
 
                             if (!reader.IsDBNull(3))
-                            { columnInfo.MaximumLength = reader.GetInt32(3); }
+                            { columnInfo.MaximumLength = reader.GetInt64(3); }
 
                             if (!reader.IsDBNull(4))
                             {
